Move to events directory
