+++ Source/CMakeLists.txt
@@ -106,6 +106,7 @@ set(libdevilutionx_SRCS
 
   lua/autocomplete.cpp
   lua/lua.cpp
+  lua/events/items.cpp
   lua/modules/audio.cpp
   lua/modules/dev.cpp
   lua/modules/dev/display.cpp
@@ -122,7 +123,6 @@ set(libdevilutionx_SRCS
   lua/modules/dev/search.cpp
   lua/modules/dev/towners.cpp
   lua/modules/i18n.cpp
-  lua/modules/items.cpp
   lua/modules/log.cpp
   lua/modules/player.cpp
   lua/modules/render.cpp
